update cyclic task tranfer
